refactor(4711): extract card-program repo ops from automation_candidates.rs giant
