+++ src/db/mod.rs
@@ -1,5 +1,6 @@
 pub mod agents;
 pub mod auto_queue;
+pub mod automation_candidate_card_program;
 pub mod automation_candidates;
 pub mod cancel_tombstones;
 pub mod dispatch_semaphores;
